+++ src/main/java/de/slikey/effectlib/util/CustomSound.java
@@ -3,6 +3,7 @@ package de.slikey.effectlib.util;
 import java.lang.reflect.Method;
 import java.util.Collection;
 import java.util.logging.Level;
+import java.util.logging.Logger;
 
 import org.apache.commons.lang.StringUtils;
 import org.bukkit.Location;
@@ -130,6 +131,10 @@ public class CustomSound {
     }
 
     public void play(Plugin plugin, Location sourceLocation) {
+        play(plugin, plugin == null ? null : plugin.getLogger(), sourceLocation);
+    }
+
+    public void play(Plugin plugin, Logger logger, Location sourceLocation) {
         if (sourceLocation == null || plugin == null) return;
 
         if (customSound != null) {
@@ -144,11 +149,13 @@ public class CustomSound {
                     Location location = player.getLocation();
                     if (location.getWorld().equals(sourceLocation.getWorld()) && location.distanceSquared(sourceLocation) <= rangeSquared) {
                         // player.playSound(sourceLocation, customSound, volume, pitch);
-                        playCustomSound(plugin, player, sourceLocation, customSound, volume, pitch);
+                        playCustomSound(logger, player, sourceLocation, customSound, volume, pitch);
                     }
                 }
             } catch (Exception ex) {
-                plugin.getLogger().warning("Failed to play custom sound: " + customSound);
+                if (logger != null) {
+                    logger.warning("Failed to play custom sound: " + customSound);
+                }
             }
         }
 
@@ -156,12 +163,18 @@ public class CustomSound {
             try {
                 sourceLocation.getWorld().playSound(sourceLocation, sound, volume, pitch);
             } catch (Exception ex) {
-                plugin.getLogger().warning("Failed to play sound: " + sound);
+                if (logger != null) {
+                    logger.warning("Failed to play sound: " + sound);
+                }
             }
         }
     }
 
     public void play(Plugin plugin, Entity entity) {
+        play(plugin, plugin == null ? null : plugin.getLogger(), entity);
+    }
+
+    public void play(Plugin plugin, Logger logger, Entity entity) {
         if (entity == null || plugin == null) return;
 
         Location sourceLocation = entity.getLocation();
@@ -174,17 +187,19 @@ public class CustomSound {
                         Location location = player.getLocation();
                         if (location.getWorld().equals(sourceLocation.getWorld()) && location.distanceSquared(sourceLocation) <= rangeSquared) {
                             // player.playSound(sourceLocation, customSound, volume, pitch);
-                            playCustomSound(plugin, player, sourceLocation, customSound, volume, pitch);
+                            playCustomSound(logger, player, sourceLocation, customSound, volume, pitch);
                         }
                     }
                 } else if (entity instanceof Player) {
                     Player player = (Player)entity;
                     // player.playSound(sourceLocation, customSound, volume, pitch);
-                    playCustomSound(plugin, player, sourceLocation, customSound, volume, pitch);
+                    playCustomSound(logger, player, sourceLocation, customSound, volume, pitch);
 
                 }
             } catch (Exception ex) {
-                plugin.getLogger().warning("Failed to play custom sound: " + customSound);
+                if (logger != null) {
+                    logger.warning("Failed to play custom sound: " + customSound);
+                }
             }
         }
 
@@ -197,7 +212,9 @@ public class CustomSound {
                     sourceLocation.getWorld().playSound(sourceLocation, sound, volume, pitch);
                 }
             } catch (Exception ex) {
-                plugin.getLogger().warning("Failed to play sound: " + sound);
+                if (logger != null) {
+                    logger.warning("Failed to play sound: " + sound);
+                }
             }
         }
     }
@@ -206,65 +223,65 @@ public class CustomSound {
         this.range = range;
     }
 
-    private static void initializeReflection(Plugin plugin) {
+    private static void initializeReflection(Logger logger) {
         if (!initializedReflection) {
             initializedReflection = true;
             try {
                 player_playCustomSoundMethod = Player.class.getMethod("playSound", Location.class, String.class, Float.TYPE, Float.TYPE);
             } catch (Exception ex) {
-                if (plugin != null) {
-                    plugin.getLogger().warning("Failed to bind to custom sound method");
+                if (logger != null) {
+                    logger.warning("Failed to bind to custom sound method");
                 }
             }
             try {
                 player_stopCustomSoundMethod = Player.class.getMethod("stopSound", String.class);
             } catch (Exception ex) {
-                if (plugin != null) {
-                    plugin.getLogger().warning("Failed to bind to stop custom sound method");
+                if (logger != null) {
+                    logger.warning("Failed to bind to stop custom sound method");
                 }
             }
             try {
                 player_stopSoundMethod = Player.class.getMethod("stopSound", Sound.class);
             } catch (Exception ex) {
-                if (plugin != null) {
-                    plugin.getLogger().warning("Failed to bind to stop sound method");
+                if (logger != null) {
+                    logger.warning("Failed to bind to stop sound method");
                 }
             }
         }
     }
 
-    public static void stopSound(Plugin plugin, Player player, String sound) {
-        initializeReflection(plugin);
+    public static void stopSound(Logger logger, Player player, String sound) {
+        initializeReflection(logger);
         if (player_stopCustomSoundMethod == null) return;
         try {
             player_stopCustomSoundMethod.invoke(player, sound);
         } catch (Exception ex) {
-            if (plugin != null) {
-                plugin.getLogger().log(Level.WARNING, "Failed to stop custom sound: " + sound, ex);
+            if (logger != null) {
+                logger.log(Level.WARNING, "Failed to stop custom sound: " + sound, ex);
             }
         }
     }
 
-    public static void stopSound(Plugin plugin, Player player, Sound sound) {
-        initializeReflection(plugin);
+    public static void stopSound(Logger logger, Player player, Sound sound) {
+        initializeReflection(logger);
         if (player_stopSoundMethod == null) return;
         try {
             player_stopSoundMethod.invoke(player, sound);
         } catch (Exception ex) {
-            if (plugin != null) {
-                plugin.getLogger().log(Level.WARNING, "Failed to stop sound: " + sound, ex);
+            if (logger != null) {
+                logger.log(Level.WARNING, "Failed to stop sound: " + sound, ex);
             }
         }
     }
 
-    public static void playCustomSound(Plugin plugin, Player player, Location location, String sound, float volume, float pitch) {
-        initializeReflection(plugin);
+    public static void playCustomSound(Logger logger, Player player, Location location, String sound, float volume, float pitch) {
+        initializeReflection(logger);
         if (player_playCustomSoundMethod == null) return;
         try {
             player_playCustomSoundMethod.invoke(player, location, sound, volume, pitch);
         } catch (Exception ex) {
-            if (plugin != null) {
-                plugin.getLogger().log(Level.WARNING, "Failed to play custom sound: " + sound, ex);
+            if (logger != null) {
+                logger.log(Level.WARNING, "Failed to play custom sound: " + sound, ex);
             }
         }
     }